requirements.txt
python-telegram-bot==21.10
torch==2.1.2
transformers==4.36.2
bitsandbytes==0.41.1
huggingface-hub==0.20.3
requests==2.31.0
python-dotenv==1.0.0
tenacity==8.2.3
accelerate==0.25.0
